+++ services/application/config/routes.php
@@ -54,7 +54,7 @@ $route["postBackend"] = "home/postBackend";
 $route["fetchAccessLevels"] = "home/fetchAccessLevels";
 $route["fetchUsers"] = "home/fetchUsers";
 $route["validateUser"] = "home/validateUser";
-$route["validateGoogleUser"] = "home/validateGoogleUser";
+$route["validateEmailProvider"] = "home/validateEmailProvider";
 $route["checkUserExists"] = "home/checkUserExists";
 $route["checkAppUserExists"] = "home/checkAppUserExists";
 $route["changePassword"] = "home/changePassword";
@@ -69,6 +69,7 @@ $route["getUniqueLocales"] = "home/getUniqueLocales";
 $route["saveLog"] = "home/saveLog";
 $route["multipleAccountsList"] = "home/multipleAccountsList";
 $route["getMultiUserRoles"] = "home/getMultiUserRoles";
+$route["auth/appleCallback"] = "appleAuth/appleCallback";
 $route["test"] = "home/test";
 
 // Account planner
@@ -123,9 +124,6 @@ $route["workbook/phpinfo"] = "workbook/workbook/phpinfo";
 
 // payment and plans
 $route["payments/availableBillingPlans"] = "payments/plans/availableBillingPlans";
-$route["payments/checkDiscounts"] = "payments/plans/checkDiscounts";
-$route["payments/checkTaxes"] = "payments/plans/checkTaxes";
-$route["payments/deductExhaustedUsage"] = "payments/plans/deductExhaustedUsage";
 $route["payments/accountClosure"] = "payments/plans/accountClosure";
 $route["payments/checkClosure"] = "payments/plans/checkClosure";
 $route["payments/revokeAccount"] = "payments/plans/revokeAccount";
@@ -141,6 +139,7 @@ $route["payments/test"] = "payments/stripe/test";
 // razorpay
 $route["payments/razorpay/test"] = "payments/razorpay/test";
 $route["payments/razorpay/createSubscription"] = "payments/razorpay/createSubscription";
+$route["payments/razorpay/createRazorpayOrder"] = "payments/razorpay/createRazorpayOrder";
 $route["payments/razorpay/onPayment"] = "payments/razorpay/onPayment";
 $route["payments/razorpay/onPaymentFailed"] = "payments/razorpay/onPaymentFailed";
 $route["payments/razorpay/onPostPaymentAutomation"] = "payments/razorpay/onPostPaymentAutomation";